+++ app/Http/Controllers/PengeluaranController.php
@@ -2,12 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\Imports\PengeluaranImport;
 use App\Pengeluaran;
 use App\ScheduleUpdate;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Crypt;
 use Illuminate\Support\Facades\Session;
+use Maatwebsite\Excel\Facades\Excel;
 
 class PengeluaranController extends Controller
 {
@@ -18,7 +20,7 @@ class PengeluaranController extends Controller
 
     public function index()
     {
-        session()->put('ibu', 'BIOS G2');
+        session()->put('ibu', 'BIOS facelift');
         session()->put('anak', 'Layanan Keuangan');
         session()->put('cucu', 'Data Pengeluaran');
 
@@ -42,28 +44,45 @@ class PengeluaranController extends Controller
         $this->validate($request, [
             'kd_akun' => 'required',
             'jumlah' => 'required',
+            'tanggal_transaksi' => 'required',
         ]);
 
-        $tanggal = Carbon::now()->format('Y-m-d');
+        $cek = Pengeluaran::where('kd_akun', $request->kd_akun)
+            ->where('tgl_transaksi', $request->tanggal_transaksi)
+            ->get();
 
-        $data = new Pengeluaran();
-        $data->kd_akun = $request->kd_akun;
-        $data->jumlah = $request->jumlah;
-        $data->tgl_transaksi = $tanggal;
-        $data->status = false;
-        // $user->akses = $request->akses;
-        $data->save();
+        if ($cek->count() > 0) {
+            Session::flash('error', 'Data pada tanggal tersebut sudah pernah dimasukkan!');
+        } else {
+            $data = new Pengeluaran();
+            $data->kd_akun = $request->kd_akun;
+            $data->jumlah = $request->jumlah;
+            $data->tgl_transaksi = $request->tanggal_transaksi;
+            $data->status = false;
+            $data->save();
 
-        Session::flash('sukses', 'Data Berhasil diperbaharui!');
+            Session::flash('sukses', 'Data Berhasil diperbaharui!');
+        }
 
         return redirect('/pengeluaran');
     }
 
     public function edit($id)
     {
+        session()->put('ibu', 'BIOS facelift');
+        session()->put('anak', 'Layanan Keuangan');
+        session()->put('cucu', 'Edit Pengeluaran');
+
         $id = Crypt::decrypt($id);
 
+        // dd($id);
+
         $data = Pengeluaran::find($id);
+        // if ($data->status == true) {
+        //     Session::flash('error', 'Data Sudah terkirim!');
+
+        //     return redirect()->back();
+        // }
         $akun = PemasukanController::RefAkun();
 
         return view('keuangan.pengeluaran_edit', compact('data', 'akun'));
@@ -72,8 +91,9 @@ class PengeluaranController extends Controller
     public function update($id, Request $request)
     {
         $data = Pengeluaran::find($id);
-        $data->kd_akun = $request->kd_akun;
+        // $data->kd_akun = $request->kd_akun;
         $data->jumlah = $request->jumlah;
+        $data->status = false;
         $data->save();
 
         Session::flash('sukses', 'Data Berhasil diperbaharui!');
@@ -85,6 +105,11 @@ class PengeluaranController extends Controller
     {
         $id = Crypt::decrypt($id);
         $delete = Pengeluaran::find($id);
+        if ($delete->status == true) {
+            Session::flash('error', 'Data Sudah terkirim, tidak bisa dihapus!');
+
+            return redirect()->back();
+        }
         $delete->delete();
 
         Session::flash('sukses', 'Data Berhasil dihapus!');
@@ -105,12 +130,12 @@ class PengeluaranController extends Controller
 
         $akun = PemasukanController::RefAkun();
 
-        return view('keuangan.pemasukan', compact('data', 'akun'));
+        return view('keuangan.pengeluaran', compact('data', 'akun'));
     }
 
     public function client()
     {
-        session()->put('ibu', 'BIOS G2');
+        session()->put('ibu', 'BIOS facelift');
         session()->put('anak', 'Layanan Keuangan');
         session()->put('cucu', 'Client Pengeluaran');
 
@@ -122,70 +147,107 @@ class PengeluaranController extends Controller
             ->whereDate('tgl_transaksi', '<', $now)
             ->get();
 
+        //Jika data pemasukan pada hari kemarin tidak ada maka input otomatis
+        $cekPemasukanKemarin = Pengeluaran::whereDate('tgl_transaksi', $now->yesterday())
+            ->get();
+        if ($cekPemasukanKemarin->count() == 0) {
+            $simpan = new Pengeluaran();
+            $simpan->tgl_transaksi = $now->yesterday();
+            $simpan->kd_akun = 525111;
+            $simpan->jumlah = 0;
+            $simpan->status = false;
+            $simpan->save();
+        }
+
         // Kirim data yang bukan hari ini
         if ($data_pengeluaran->count() > 0) {
-            // dd($data);
-            $schedule = ScheduleUpdate::all();
-            foreach ($schedule as $jadwal) {
-                if (($jam >= $jadwal->waktu_mulai) and ($jam <= $jadwal->waktu_selesai)) {
-                    foreach ($data_pengeluaran as $data) {
+            // dd($data_pengeluaran);
+            // $schedule = ScheduleUpdate::all();
+            // foreach ($schedule as $jadwal) {
+            //     if (($jam >= $jadwal->waktu_mulai) and ($jam <= $jadwal->waktu_selesai)) {
+            foreach ($data_pengeluaran as $data) {
 
-                        $tgl_transaksi = Carbon::parse($data->tgl_transaksi)->format('Y/m/d');
-                        $client = new \GuzzleHttp\Client(['base_uri' => session('base_url')]);
-                        $response = $client->request('POST', 'ws/pengeluaran/prod', [
-                            'headers' => [
-                                'token' => session('token'),
-                            ],
-                            'form_params' => [
-                                'kd_akun' => $data->kd_akun,
-                                'jumlah' => $data->jumlah,
-                                'tgl_transaksi' => $tgl_transaksi,
-                            ]
-                        ]);
+                // $tgl_transaksi = Carbon::parse($data->tgl_transaksi)->format('Y/m/d');
+                $client = new \GuzzleHttp\Client(['base_uri' => session('base_url_bios')]);
+                $response = $client->request('POST', 'ws/keuangan/akuntansi/pengeluaran', [
+                    'headers' => [
+                        'token' => session('tokenBios'),
+                    ],
+                    'form_params' => [
+                        'kd_akun' => $data->kd_akun,
+                        'jumlah' => $data->jumlah,
+                        'tgl_transaksi' => $data->tgl_transaksi,
+                    ]
+                ]);
 
-                        $datajson = json_decode($response->getBody());
+                if ($response->getStatusCode() == 200) {
 
-                        // dd($data);
+                    $datajson = json_decode($response->getBody());
 
-                        if ($data->status != 'MSG20003') {
-                            Session::flash('error', $datajson->message);
+                    // dd($datajson, $data);
 
-                            return view('keuangan.client_pemasukan');
-                        } else {
-                            Session::flash('sukses', "$datajson->status, $datajson->message");
+                    if ($data->status != 'MSG20003') {
+                        Session::flash('error', $datajson->message);
 
-                            $update = Pengeluaran::find($data->id);
-                            $update->status = true;
-                            $update->save();
-                        }
+                        return view('keuangan.client_pemasukan');
+                    } else {
+                        Session::flash('sukses', "$datajson->status, $datajson->message");
+
+                        $update = Pengeluaran::find($data->id);
+                        $update->status = 1;
+                        $update->save();
                     }
+                } else {
+                    Session::flash('error', 'Pengiriman gagal status pengiriman ' . $response->getStatusCode());
+                    goto Selesai;
                 }
             }
+            //     }
+            // }
         }
+
+        Selesai:
         $tgl_mulai = new Carbon('first day of this month');
         $tgl_selesai = new Carbon('last day of this month');
 
         $data = Pengeluaran::whereDate('tgl_transaksi', '>=', $tgl_mulai)
             ->whereDate('tgl_transaksi', '<=', $tgl_selesai)
+            ->orderBy('tgl_transaksi', 'DESC')
             ->get();
 
-        $kemarin = Carbon::now()->yesterday()->format('Y/m/d');
-        // $kemarin = "2022/03/25";
-        $client = new \GuzzleHttp\Client(['base_uri' => session('base_url')]);
-        $response = $client->request('POST', 'get/data/pengeluaran', [
-            'headers' => [
-                'token' => session('token'),
-            ],
-            'form_params' => [
-                'tgl_transaksi' => $kemarin,
-            ]
+        return view('keuangan.client_pengeluaran', compact('data'));
+    }
+
+    public function import(Request $request)
+    {
+        // validasi
+        $this->validate($request, [
+            'file' => 'required|mimes:csv,xls,xlsx'
         ]);
 
-        $datajson = json_decode($response->getBody());
-        $cekdata = $datajson->data;
+        // menangkap file excel
+        $file = $request->file('file');
 
-        // dd($cekdata);
+        // dd($file);
 
-        return view('keuangan.client_pengeluaran', compact('data', 'cekdata'));
+        // membuat nama file unik
+        $nama_file = rand() . $file->getClientOriginalName();
+
+        // upload ke folder file_siswa di dalam folder public
+        $file->move('bios/pengeluaran', $nama_file);
+
+        try {
+            // import data
+            Excel::import(new PengeluaranImport, public_path('/bios/pengeluaran/' . $nama_file));
+
+            // notifikasi dengan session
+            Session::flash('sukses', 'Data Berhasil Diimport!');
+        } catch (\Throwable $th) {
+            // notifikasi dengan session
+            Session::flash('error', 'Cek kembali data file Anda!');
+        }
+
+        // alihkan halaman kembali
+        return redirect()->back();
     }
 }
